workers/requirements.txt
playwright==1.44.0
psycopg2-binary==2.9.9
apscheduler==3.10.4
python-dotenv==1.0.1
playwright-stealth==1.0.6
httpx==0.27.0
pydantic==2.7.4
loguru==0.7.2
tenacity==8.3.0
